IoT_Database/requirements.txt
certifi
chardet
dataclasses
Django
django-stubs
djangorestframework
djangorestframework-stubs
idna
mongoengine
mypy
mypy-extensions
pymongo
pytz
requests
six
typed-ast
typing-extensions
remote-pdb
pyjwt
passlib
